requirements.txt
fastapi
uvicorn[standard]
websockets
paho-mqtt
